Fix numerous thinkos confusing 'build' vs 'binary'
What's annoying is this worked on Launchpad i386 but not Launchpad
amd64...however, it'd pdebuild just fine.

Eventually diffing the two outputs and comparing them carefully, we
see the following in the broken build:

  /usr/bin/fakeroot debian/rules binary-arch

And this in the working build:

  /usr/bin/fakeroot debian/rules binary

I then realized that binary-arch was doing what it was supposed to do
and looked more closely, and realized I had stuffed everything up but
had not noticed.

Signed-off-by: Daniel Farina <daniel@heroku.com>